tracer: add Tracer fn for processing a Payload directly
Users may not receive a `decoder::payload::InstructionTrace` directly
from decoded packets but a `decoder::payload::Payload`. For them, it may
be awkward to match the payload. Thus, we decided to allow feeding these
`decoder::payload::Payload`s directly, with a name that is also more
obvious than the `process_te_inst` we have for historical reasons.