examples: migrate simple example to ephemeral SMI packets
We plan to make `smi::Packets` ephemeral, allowing users to decide
whether or not to decode its payload. With the new interface, the data
members of `smi::Packets` will not be public any more.

We already introduced fns that make up part of the new interface. This
change migrates the example which accessed data members directly.