decoder: make smi::Packet data members private
We plan to make `smi::Packets` ephemeral, allowing users to decide
whether or not to decode its payload. It will carry both a decoder and
reset data, as _private_ data members. We don't want to mix public and
private data members in a struct. Therefore, we need to make the
existing data members private.