decoder: add smi::Packet fns for querying the trace type
We recently made `smi::Packet` ephemeral, allowing to decode the payload
only when it is a trace payload we deem relevant and are able to decode.
However, the decision whether or not to decode a payload may need to be
made based on the trace type. Previously, we'd lack interfaces for
querying that value.