decoder: remove Decoder::decode_restricted
We originally introduced the fn to facilitate `Decode` impls for packet
types such as `smi::Packet`. However, we since switched to ephemeral
packet types that make use of `Decoder::split_data` and `Decoder::reset`
directly.

And unlike `Decoder::decode_restricted`, the new packet decoding scheme
leaves the `Decoder` at the border to the next packet even if decoding
of a packet's payload fails.