examples: make simple example print bytes left on separate line
As part of debug output, the simple example prints every decoded packet
and the number of bytes of trace data left to decode. However, when we
migrated to ephemeral packets in

    a6bf858894d67959cd671f35034c10c8947bec02

we needed to mode retrieval of that value before the decoding. Thus, the
value displayed now includes the size of the packet, while still
indicating something else.

This change resolves this issue by moving the number of bytes still to
decode to a separate statement which is printed when debug is enabled.