examples: make simple example print decoded payload
Prior to the transition to ephemeral packets in

    a6bf858894d67959cd671f35034c10c8947bec02

the simple example's debug output would include the decoded payload —as
part of the decoded packet. Since the transition, the payload is not
decoded eagerly, and the `Debug` impl for `decoder::smi::Packet`
represents the payload as the byte slice it is at that point. Still, we
want to see what got decoded.