instruction: impl info::Info for Kind
We plan to provide support for user-defined instruction sets, e.g. for
support of ISA extensions. Apart from the decoding, users may want to
specify their own instruction `Kind`s, which implies a user-provided
type. For this purpose, we introduced a new trait `info::Info` as an
abstraction defining the query fns we need for the core tracing logic.

This change implements that trait for `Kind`. We preserve the existing
interface for now. We'll remove the old impl block once we migrated all
existing users.