instruction: remove old query fns associated to Kind directly
We plan to provide support for user-defined instruction sets, e.g. for
support of ISA extensions. Apart from the decoding, users may want to
specify their own instruction `Kind`s, which implies a user-provided
type.

The core tracing logic relies on a collection of query fns associated to
`Kind`. These changes move these fns into a trait which is implemented
for `Kind`, but also for `Option` and `Instruction`.