binary: remove constraints on boxed::BoxedError's type parameters
when introducing `boxed::BoxedError`, we added type constraints on its
`B` parameter not only to the `Binary` impl for which the constraints
were necessary but the type itself. This had the benefit that you could
not create a `boxed::BoxedError` that would not function as a `Binary`.

However, we plan making `Binary` generic over `instruction::info::Info`.
Constraining that paramter for `boxed::BoxedError` would add even more
noise and pose an unnecessary restriction —it would not allow deciding
which `Info` to use after creation.