instruction: add trait for decoding an info::Info
We recently introduced a trait abstracting instruction information
required for tracing. This was motivated partially by the prospect of
generic `Instruction`s that would allow user-supplied types as a
replacement for `Kind`.

However, we did not yet address instruction decode. Currently, this is
done through fns associated to `base::Set`. This fn introduces a trait
abstracting the decoding, mainly mirroring the signatures of those
existing associated fns.

We chose a generic trait over an associated `Info` type as this allows,
for example, to implement decoding of several different `info::Info`
types for `base::Set`.