instruction: impl info::Decode<Option<Kind>> for base::Set
We recently introduced a trait `info::Info` abstracting instruction
information required for tracing. This was motivated partially by the
prospect of generic `Instruction`s that would allow user provided types
as a replacement for `Kind`. We also introduced a trait for decoding
these `info::Info` types. This change now adapts `bae::Set` so that it
implmenets that trait rather than just having a bunch of associated fns.