tracer: move Builder's binary::Binary type constraint to build fn
We recently introduced a type parameter to `Tracer` that allows to
specify an `instruction::info::Info` to use for tracing —boasting that
we had removed the last obstacle preventing user to use their own
types. However, we do need the ability to construct those `Tracer`s, and
up until now `Builder` sould still only build `Tracer`s with the default
`Option<instruction::Kind>`.

This change resolves the issue by removing the `Binary` type contraint
from all but `Builder::build` and adapting the constraints for that fn
such that it did not unnecessarily contraint the instruction item type.