instruction: fixed error in fn decode_16()
Previously we wanted to ensure that compressed instruction is c.lui if
the destination register is neither 0 nor 2, as shown in Chapter 27.5.1.
Integer Constant-Generation Instructions of RISC-V Manual of
Unprivileged Architecture. However the checking for this condition XOR
was used which results in true in case of rd = 0 or rd=2 as one of both
conditions is always true. Therefore XOR is replaced by AND.