instruction: impl TryFrom<{u16,u32,u64}> for bits::Bits
We may sometimes want to create `bits::Bits` from an `u16` or `u32` and
rely on type inference to figure out the correct variant. However,
`From` impls for `u16`, `u32` and `u64` would probably end up being
foot-guns in practice while also leaving out 48bit instructions.

What we really want is something a bit smarter: check whether an integer
conforms to the length encoding for instructions of that integer's
length. And while we're at it, also convert to shorter instruction
`bits::Bits` if applicable.

This change introduces `TryFrom` impls doing just that. We end up
dumplicating some logic for differentiating instruction lengths, which
is not ideal but still acceptable.