instruction: add trait for creating info::Decode for RV32I and RV64I
We allow for user provided `info::Decode` implementations and provide an
impl for `riscv_isa::Target` ourselves. Some `binary::Binary`s may need
to construct an `info::Decode` internally, e.g. because we don't want to
burden the user with deriving it from some piece of information.

Previously, we would only do this for `base::Set`. This change
introduces a trait that allows type-agnostic construction of some
`info::Decode`.