tracer: add optional PC to IterationState::ContextItem
Our current reporting of an EPC is faulty for some double traps. The fix
requires resetting the current address in the `Tracer`'s `state::State`
in some cases. At the same time, however, we want to emit a context item
with a _different_ PC, namely the address that was reported for the
trap.

In order to have both, we need to record that PC in the `IterationState`
and retrieve it when emitting the respective items. The variant for
traps already has a member for the EPC which we can use. This change
adds a member to the variant for context items. Since context items are
not only emitted for traps, we make the member an `Option`.