tracer: reset state to trap address if thaddr is unset
Our current reporting of an EPC is faulty for some double traps. We
would compute the EPC baed on the current `state::State` and not reset
the state for cases where the instruction before the trap was reported.
For the second trap, our EPC would therefore be based on incorrect
state.

This change resolves this particular issue. EPC computation, however,
remains faulty if the second trap is an exception. That issue will be
addressed by a future change.