tracer: reset state on start payload after an exception
Our current reporting of an EPC is faulty for some double traps. The
fix involves changes to the computation of the EPC. This would break the
behaviour for single traps if the exception or interrupt occurs after an
uninferable PC discontinuity. In such cases, the encoder emits a trap
packet with `thaddr == 0`, followed by a start packet at the trap
handler's position.

Fortuantely, that issue can be fixed by adding additional responsibility
for resetting the traced address to the code handling start packets.
Interrupts are sufficiently similar to the conditions at the beginning
of a trace.