tracer: do not try to compute a new EPC after an exception
Our reporting of an EPC was faulty for some double traps. The EPC for
exceptions is normally calculated by advancing the current state. For an
exception right after a trap, however, the state was reset to an address
of the instruction that did cause the exception and thus not retire. The
EPC is therefore the current address.