tracer: fix branch value emitted by tests::start_packet
According to the specification, the `branch` field of a start packet is
`0` only if `address` points to a branch instruction and the branch was
taken. In any other case the field is supposed to be `1`. As the data
member `start::Start::brnach` represents that bit one-to-one, it should
also be `true` for a regular start packet for a trace not starting at a
branch.