decoder: make tests::bitstream_test! accept complete config::Parameters
When introducing the `tests::bitstream_test!` macro, we did so with the
intention to rewrite a number of existing tests, some of which set some
individual config parameters. In those circumstances, we did not want to
build a complete `config::Parameters` but inject only the non-default
parameters.

For tests derived from the reference flow, however, we'll add quite a
few tests with an identical, non-default `config::Parameters`. In that
situation, we _do_ want to pass a reference to a complete struct
provided through a common facility.