decoder: added tests for decoding instruction trace
In order to test out if payload of data is correctly decoded test were
written for synchronization and branch type of an instruction trace.
Currently decoder parameters are given in every time in code format and
have to be defined for every test. Idea in future is to parse it in as
some constant in order to not need to redefine it every time