tracer: add testing binary for testing address payload details
Section 7.6.2 of the specification defines the specifics around the
`notify` and `updiscon` fields in branch and address payloads (which
both effectively map to `decoder::payload::AddressInfo`). These fields
cover edge-cases which require certain, rare structures in traced
binaries. The section describes such a structure.

We want to test our tracer in those edge-cases, and need a binary that
contains a structure which allows for them to occur.