tracer: remember exception via Option of an enum dedicated to events
We previously introduced a `Tracer` data member indicating whether the
previous packet we processed expressed a specific condition. We need
this piece of information for properly processing double traps.

We have similar needs for other parts of the overall tracing logic. We
decided to track all of these events through a common mechanism rather
then adding more and more specialized flags (that also need to be reset
on certain occasions).