tracer: remember previous address packet
Section 7.6.2 of the specification defines the specifics around the
`notify` and `updiscon` fields in branch and address payloads —which
both effectively map to `decoder::payload::AddressInfo`. Handling these
edge-cases requires deferring decisions until the the processing of the
packet _after_ certain `decoder::payload::AddressInfo`. We thus need to
record that we processed an `decoder::payload::AddressInfo`.