tracer: add state::State fn for querying the previous Instruction
Some tracing logic for processing of `decoder::payload::AddressIndo`
needs to distinguish between whether the `item::Item`s derived from a
previous `decoder::payload::AddressIndo` ended with an uninferable PC
discontinuity or not. In order to check this additional condition, we
need to access the previous instruction recorded in the `state::State`.