tracer: make state::State inferred on consequtive addr/branch payloads
Section 7.6.2 of the specification defines the specifics around the
`notify` and `updiscon` fields in branch and address payloads —which
both effectively map to `decoder::payload::AddressInfo`. We previously
failed to handle `decoder::payload::AddressInfo`s (with `notify` and
`updiscon` set to false) followed by another address or branch payload
correctly.

A trace encoder will issue an `decoder::payload::AddressInfo` only if an
address must be reported, i.e. for uninferable PC discontinuities and a
few other cases such as explicit notification (`notify` flag set). In a
certain edge case (case 1 described in section 7.6.2), an `AddressInfo`
issued due to a discontinuity is undistinguishable from one issued
because of a trap or other event. It is only with the next packet that
we can distinguish between these cases.

If the next packet is a branch or address packet, the previous packet
cannot have been issued due to a special event. Thus, we derived too few
`item::Item`s from the previous `decode::payload::AddressInfo`. We can
make up for that by switching the `state::State` into inferred mode,
This will cause the generation of `item::Item`s up to the dicontinuity
which caused the previous `decode::payload::AddressInfo` to be issued,
folloed by processing of the second packet.