tracer: add data member for address width to state::State
We plan to restrict traced PCs to the address width defined by the
encoder parameters. As we need to do so after computing an PC but before
querying an `instruction::Instruction` for that PC, we need to do so in
`state::State::next_pc`. Thus, the width needs to be availible in that
fn, and we rather have a new data member than another parameter. In the
future, it will replace `Tracer::address_delta_width`.