tracer: move responsibility for sign-extending address to state::State
Previously, the address width was stored the `Tracer` for the purpose of
sign-extending addresses extracted from `decoder::payload::AddressInfo`
only. However, we recently added another data member holding the address
width to `state::State` for the purpose of restraining computed PCs.

Storing this value in both `Tracer` and `state::State` is a waste of
resources. Given that is used exclusively for computing an input to
`state::Initilizer::set_rel_address`, we decided to shift responsibility
for performing this computation away from `Tracer`, allowing the removal
of `Tracer::address_delta_width` in the future.