tracer: enable specifying encoder parameters in trace_test!
The `trace_test!` was designed tailored to the test-cases present at the
time, with the default `Tracer` configuration. However, we plan to
introduce some tests that require specific configuration of the `Tracer`
in the future. This change introduces the basic extension mechanism as
well as means to supply alternative parameters.