tracer: restructure test generation macros
The previous structure of the test generation macros demanded more
repetition that we would like, and to prevented some extensions for the
generation of multiple tests.

This change resolves the issue by offloading the expansion to a helper
macro, using constructs that can be handled and forwarded as a single
token tree that is only mached against a more elaborate construct when
necessary.