tracer: add basic "jumps" test for implicit return behaviour
The implicit return option is known to be broken in the reference flow's
encoder. We therefore cannot verify the many test-cases we would want to
write. This test-case is, however, simple enough that both the reference
flow does the right thing, i.e. we can reason about the trace it
produces is correct.