tracer: make implicit return option settable via state::State::new
We want the ability to set supported options when creating a `Tracer`.
This setting is stored in the `state::State` of a tracer and can be set
via a `state::Initializer`. However, this process is a bit awkward and
we already have precedence of setting the sequentially inferable jumps
setting via `state::State::new`. We therefore decided to take the same
appraoch with the implicit return setting.