tracer: remove special v1 support logic from Tracer
Originally, the tracer was essentially transcribed from pseudeo-code
found in version 1 of the trace specification and later adapted to
version 2, with the intention of preserving support for version 1.
Supposed differences in the tracing algorithm were handled via a
feature.

When we set out to reshape `Tracer` into what it is today, we took care
to preserve the logic for supporting version 1. However, did not get
around to actually validate version 1 support until recently. When we
checked the specification and the reference implementation, there
appears to be no mismatch between version 1.1.3 and 2.0.3 that would
warrant the distinction —aside from the pseudeo-code which we did not
check in detail.

Since the reference implementation did not change, we assume that the
version 1 specific code does actually not provide functionality the
version 2 cases do. On the contrary, the version 1 cases are likely to
cause issues. We therefore decided to remove this logic and make the
version setting for `Tracer` a dummy setting for the time being.