tests: make tracer::tests a toplevel module
We plan to introduce trace encoding functionality in the future, and we
would like to use the existing tests for that —by extending the test
generation macros.

Trace encoding will not live in the `tracer` but in some othe rmodule
we'll introduce then. This would make the previous placement of those
tests awkward. Making `tracer::tests` solved that issue and seems
appropriate given that the functionality tested is this library's core
functionality.