packet: rename decoder module to packet
We plan to introduce encoding capabilities in the future. The
implementation would look similar to `Decoder` and `Decode`, and we
could just add an extra `encode` module. However, we would want to impl
an `Encode` trait (primarily) for payloads which live in `decoder`.
Which is awkward. And hading the `Enocder` be part of `decoder` is just
as awkward.

Since the existing module holds all these payload and packet
definitions, we decided that `packet` would be just as fitting as
`decoder` anyway.