decoder: make payload::AddressInfo::address an i64
The default mode of operation is delta mode, in which the `address` in
`payload::AddressInfo` denotes an address relative to the last reported
one. Still, we previously decided to use an `u64` for that data member
and thus had to sign-extend the value in the tracer.

This change makes `payload::AddressInfo::address` an `i64` , allowing to
just add it to the last reported address (in the future) since it is
already sign-extended during decode. Meanwhile, absolute 32bit addresses
in the range starting from `0x80000000` are still handled correctly by
the tracer thanks to the existing masking code in
`tracer::state::State::next_pc`.