packet: move Decocer and Decode to dedicated submodule
We plan to introduce encoding capabilities in the future. This does
include encoding of payloads and packets —which motivated the recent
renaming of `decoder` to `packet`. We don't want to sqeese both the
alreagy lengthly `Decoder`/`Decode` and the future encoder in a single
file. We therefore decided to move `Decoder` and `Decode` into a
dedicated submodule and to do the same for a future `Encoder`.

The `Builder` will remain in `packet` itself as we'll also use it for
constructing `Encoder`s.