packet: add Encoder type and Encode trait in dedicated encoder module
We want to support not only decoding packets and payloads but also
encoding them. For that we decided to mainly mirror the structures we
have for decoding: a type carrying necessary state as well as
configuration and a trait that specifies how a specific payload or
component is handled.

One detail that does differ from the existign decoding infrastructure is
that we choose to make the latter a public trait, allowing for a simple
`Endoder::encode` fn rather than a set of fns, each for a different
type.