packet: add Builder fns for building encoder::Encoders
`Encoder` is, by design, very similar to `Decoder` both in terms of its
structure and interface as and in what data it cerries. Thus, it only
makes sense to construct it the same way. And instead of introducing a
new `Builder` just for encoders, we may as well re-use the existing one.