packet: add tests submodule for basic encoder tests
We have quite a few tests in the `packet` module. Enough that we want to
group some of them into submodules. For the basic tests, howver, we want
to use this opportunity to convert them into tests generated through a
macro. This is slightly more concise, but more importantly allows us to
re-use the test cases for encoding tests in the future.