packet: add type for scoping a decoder::Decoder
Currently, both `smi::Packet` and `encap::Normal` encapsulate a
restricted `decoder::decoder` and the "remaining" data after that. And
both reset the `decoder::Decoder` to that "remaining" data when dropped.

This change introduces a type that will allow us to eliminate this
duplication and offers opportunities for derived `Debug` and `PartialEq`
as well as easier genericalization of those packet types in the future.
The latter will eventually allow for packets that are either partially
or fully decoded, i.e. contain a `payload::Payload`.