packet: use decoder::Scoped for restricting decoder in encap::Normal
When we introduced `encap::Normal` in

    705d12662e6c89e1a7100be440ab94842f43946e

we decided to not do that via a type abstracting the scoping but
handling of the `decoder::Decoder` direclty in `encap::Normal`. This
led to some code duplication w.r.t. `smi` and now poses an obstacle for
introduction of non-ephemeral packets that we actually _do_ want to
encoding.

We recently introduced `decoder::Scoped` with the goal to centralize the
scoping for ephemeral packets. This change starts the migration for
`encap::Normal`, eliminating some hand-rolled trait impls for the type.