packet: use decoder::Scoped for restricting decoder in smi::Packet
When we made `smi::Packet`s ephemeral in

    0426ccafd91d79d9e603ee7b4288859850bea1f9

we decided to not do that via a type abstracting the scoping but
handling of the `decoder::Decoder` direclty in `smi::Packet`. This led
to some code duplication w.r.t. `encap` and now poses an obstacle for
introduction of non-ephemeral packets that we actually _do_ want to
encoding.

We recently introduced `decoder::Scoped` with the goal to centralize the
scoping for ephemeral packets. This change starts the migration for
`smi::Packet`, eliminating some hand-rolled trait impls for the type.