packet: rename smi::Packet::payload to decode_payload
We recently made `smi::Packet` generic over its payload, allowing for
non-ephemeral `smi::Packet`s. for those, we'll want the ability to
access the payload, and the most suitable name for a fn exposing it
would be `payload` —which clashes with the existing fn.

Since that fn does not just expose a payload but decodes it, we decided
to rename it accordingly.