packet: impl encoder::Encode for smi::Packet
This impl allows encoding `smi::Packet`s if the payload implements
`encode::Encode`. We do intend to impl that trait for different payload
types, including `payload::InstructionTrace`. However, we do not intend
to implement it for `payload::Payload`. It will be the users
responsibility to assemble an `smi::Packet` with the appropriate type
and `trace_type` for encoding.