packet: remove Builder::default_encoder
We recently introduced `encoder::Encoder` for encoding packets and
payloads. The current design has `encoder::Encoder` contann a generic
buffer as an `AsMut<u8>`. Unfortunately, this choice obstructs the
introduction of facilities for sub-division, which are necessary for
encoding framed packet formats. We therefore decided to transition to a
concrete `&mut [u8]`. Its default value does not make much sense for
encoders.