packet: make encoder::Encoder carry a &mut [u8] as data
We recently introduced `encoder::Encoder` for encoding packets and
payloads. The original design had `encoder::Encoder` contain a generic
buffer as an `AsMut<u8>`. That worked well enough for encoding payloads.

When encoding framed packets, however, we need to either write to parts
of the buffer that are already committed or reserve space for packet
headers. At the same time we need to reset the compression state such
that compression does not cross packet boundaries.

An interface that would allow access to already-committed parts of the
buffer, paired with an interface for managing the commit state would
allow for such, but would also be quite fragile. Instead, we decided to
follow a similar approach as with `decoder::Decoder` and make use of
`slice::split_*` and similar fns for a much more rigourous interface.

Unfortunately, such an interface is obstructed by our original choice of
a generic buffer. The sub-division unterface we intend to introduce
requires the ability to reset an `encoder::Encoder` to a new buffer,
which needs to be of the same type. The aforementioned `slice` fns
produce slices, which are not convertible to arbitrary `AsMut` types. We
could introduce additional type constraints, but the most sane solution
is to just have the data member a `&mut [u8]`.