packet: add encoder::Encoder fn for resetting internal data
We decided to provide an interface for dividing the buffer written to,
in order to aid with the encoding of framed packet formats. This does
also require updating/resetting the buffer. As this also requires
resetting some other state, we decided to make this a dedicated fn
similar to `decoder::Decoder::reset`.