packet: add encoder::Encoder fn for reserving a chunk in the buffer
In most packet formats, the header containing the packet's length
preceeds the payload. Since we know the length only after encoding, we
needs the ability to reserve space for the header. This change
introduces a fn for achieving this by dividing the `encoder::Endoder`'s
buffer, and thus extract a mutable refrence to the region covering the
packet's header.