packet: add encoder::Encoder fn for determining uncommitted bytes
We are encoding items into some underlying, pre-allocated buffer.
Naturally, users of the `encoder::Endoder` will want to transmit only
the portion of that buffer that was written or, to be more precise,
comitted. Naïvely, this would be `bytes_comitted`. However, encoding
packets, and potentially other entities, will result in the buffer being
reset to a sub-region of the original buffer.

Fortunately, the `encoder::Encoder` will always be reset to the last
part of the original buffer, i.e. the end of the original buffer is
identical with the end of the new buffer. Thus, we can always determine
the comitted region from the original buffers' length and the current
number of uncomitted bytes.