packet: add new generation macro for general tests in new submodule
We currently generate some tests using the `tests::bitstream_test!`
macro, which we recently extended for encoder tests. That macro
currently allows specifying `config::Parameters`, but for some packet
formats we may want to define supplemental widths.

Rather than adding new parameters we decided to pursue a similar
approach as with trace tests. Fortunately, we only have to consider a
single, isolated packet per test. This allows for a more simple
implementation.

We could change the existing macro and all affected users at the same
time in a single commit, but we decided to take this opportunity and
put the (new) macro in a dedicated submodule. The new macro is shadowed
by the old one until we remove it.