packet: fix handling of encap srcID widths that are not a multiple of 8
Previously, we would "cut" the decoding buffer just before the payload,
after decoding the srcID and timestamp fields. While the single-byte
header is always byte-aligned, the payload may not be byte-aligned since
the srcID's width may be an arbitrary number of bits wide. Thus, cutting
the buffer would only work if srcID's bit-width was a multiple of eight.

This change resolves the issue by cutting right after the header,
calculating the length by adding the number of full bytes of the srcID
and the timestmp-width to the length field extracted from the header.