types: move tracer::stack module to types
We plan to introduce utilties for generating tracing packets with the
goal to allow simulation of a tracing unit. For one particular variant
of the implicit return feature, the tracing unit/generator needs to keep
a stack of return addresses, just like the tracker. Naturally, we would
like to re-use the utilities we introduced for that as part of `tracer`.

Since we want to avoid a dependency of a future `generator` module on
`tracer`, we decided to move `stack` to another module in preparation
for its future use.