config: add type representing optional feature selection
Currently, we manage optional features via discrete flags in
`tracer::state::State`. We would need to, more or less, duplicate this
approach in a future packet/payload generator, including the extraction
of features from `Parameters` and `packet::unit::IOptions`.

Introducing a new type for the features (which we support) allows moving
the extraction to common code, and may also offer opportunities for
communicating features to users in the future.