generator: add enum representation of itype defined by the spec
The specification defines an interface between HART and trace encoder.
The most central type of that interface is the `itype`, which is
described as either a three or four bit wide signal. Those variants
differ with regard to the categorization of jumps. While the three bit
variant can only express uninferable jumps, the other allows
differentiation between different kinds of inferable and uninferable
jumps. The differentiation is required for some optional trace features.

We considered introducing type-level differentiation between `itype3`
and `itype4`, but deemed the benefits not worth the additional
complexity.