generator: add initial Generator implementation
We previously introduced some utilities for payload generation as part
of the `generator` module. Up until now we still lacked an actual
implementation for the trace (encoder) algorithm. This change introduces
such an implementation, as part of a new `Generator` type.

Users do not yet have the ability to drive the implementation. Future
changes will introduce interfaces for that.