generator: add a trait for trace data emitted within a single exec step
The trace encoding we plan to implement takes as input data concerning a
few consequitive states of execution. In a hardware implementation,
these are signals usually following the hart to encoder interface.
However, in instruction set simulators and emulators, the availible data
may differ somewhat, and a detour via an something closely modelled
after a hardware interface may not be the most sensible option.

We therefore decided to introduce a trait as an interface that is more
abstract than a signal-by-signal hart to encoder interface and at the
same time more focussed on the needs of the encoder.