packet: make sync::Context::context an u64 rather than Option
As its name suggests, `packet::sync::Context::context` is a data field
meant for distinguishing different tracing contexts. It may not be
included, based on the trace encoder parameters. We thus decided to make
it an `Option`. However, we since regretted this decision: whether or
not it is included is _static_ for a given trace, users will expect its
presence or absence and its opaque to us anyway.

As we deem `0` to be an acceptable value for an absent context field, we
decided to make it a plain `u64` after all.