generator: add module for future payload generators
This library allows decoding trace packets and payloads as well as
tracing programs based on those payloads. In addition, we recently added
support for encoding packets and payloads with the goal to allow
generation of traces.

We currently still lack utilities for generating payloads based on
suitable tracing inputs. We plan to introduce these in the future, in
the separate module introduced by this change.