generator: add a dedicated error type for generators
In theory, it is possible to build an infallible trace encoder. Contrary
to a decoder, an encoder only requires a single data source providing
inputs for which we can easily model all valid (however nonsensical)
values using relatively simple types.

However, we plan to extract features from `ioptions` supplied by the
user —just like we do for `tracer`. Since that is a fallible operation,
we need to report those errors to the user. We also plan to provide not
only an hart-to-encoder-like interface but also more low-level
interfaces for issuing packets. Those interfaces will introduce fallible
operations.