generator: add fn for processing a new cycle
We recently introduced an implementation of the trace payload generation
algorithm. However, we were lacking means for users to interact with the
implementation as it is a private fn. We chose to do so because we
wanted a more strict public interface that would, for example, not allow
feeding events without a step.

This change introduces that interface.