generator: add Generator fn for beginning qualification
Trace payloads are only generated for instructions that are qualified.
These may be all instructions or only instructions in a specific address
range or following the execution of a specific instruction. Before
sending the first "real" payload containing a (qualified) address, we
need to issue a support payload containing `ioptions`.

This change introduces a fn for generating such a payload and
configuring the `Generator`'s feature selection according to the
`ioptions`.