generator: add Generator fn for ending qualification
When we ran out of `step::Step`s to trace for one specific instance, we
need to drain the remaining current `step::Step` and issue a support
payload declaring the end of qualification. This change introduces a fn
for ending qualification as well as an `Iterator` which returns the
outstanding payloads.