tests: differentiate repetitions better in trace_test!
Some tests involve repetitions (e.g. of loop iterations) for which we
introduced syntax within our test generation macros. When we did we
simply used paranthesis as we did for the items. This lead to a clash
that did not manifest only due to the number of token-trees/expressions
we ended up wtih in our test definitions.