tests: make trace_test! accept hints in items
We plan to introduce trace encoding functionality in the near future.
For testing that functionality, we intend to use existing test-cases.
However, some aspects such as (non)taken branches before an exception
are expressed via packets but not visible in the items from which we
will derive the generator input. We therefore need to introduce
additional hints.

This change allows placing such hints without breaking the existing
generation of `decoder` and `size_hint` tests.