tests: add type for hints attached to items
We plan to introduce trace encoding functionality in the near future.
For testing that functionality, we intend to use existing test-cases.
However, some aspects such as (non)taken branches before an exception
are expressed via packets but not visible in the items from which we
will derive the generator input. We therefore need to introduce
additional hints.

This change introduces a type we may use as a helper for accumulating
and a container holding some hints we may need in future payload
generation tests.