tests: extend full_branch_map test to accomodate for encode tests
We plan to introduce generator capabilities to this libraries in the
future, and we intend to use our existing test cases for the encoder.
That only works if the last packet is valid for an end of qualification.
This is not the case for a packet reporting a full branch map.