tests: annotate sync events
Sync events are not inferable from the code or tracing items. Instead,
they are usually controlled via a counter in or near a trace encoder.
Since in our tests we are not interested in those counters but in the
behaviour if a sync event happens, we placed them arbitrarily.

This change adds hints for those events, allowing reconstruction of the
modelled behaviour in future payload generation tests.