generator: assume emission of sync packet only if options are set
After an updiscon, we need to report either the current address or a
trap if that's the current `step::Kind`. If we report an address, we
need to signal `state::Reason::Updiscon` if the next step would result
in the emission of a sync payload. This would usually be the case when
we reach the end of qualification. However, we don't enforce all aspects
of qualification in our interface. And we make use of that in some of
our tests that are based on "snippets" of a trace, i.e. qualification
aspects are not capured.

The final support payload is only emitted if the `options` are set,
usually through `Generator::begin_qualification`. I.e. we only model the
end of qualification in its entirety if those are not `None`.