generator: add fn for creating a step::Step from a retired instruction
In both trace encoder plugins for instruction set simulators or
emulators and our very own tests, we have traps and retired instructions
as separate events. For the latter, we'll often not have the branch
information for the cycle itself but only as an observable effect via
the following instruction. The code necessary to compute the
`step::Step`s from that will be very similar among implementations.

At least for cases in which the instrucitons are availible, or easily
transformable, into an `instruciton::Instruction`, we can at least
provide some convenience functionality for determining the `step::Kind`.