generator: add fn for retrieving the instruction size in a step::Kind
The size of retired instructions is part of `step::Kind` variants since
(a) we only care bout that size in specific circumstances and (b) it may
not be present for traps but is furanteed to be present for all other
`step::Kind`s. However, other users may want more convenient access to
this value.