doc: replace test-reference.sh script with a Makefile
For testing against traces generated by the reference flow, we
previously used a shell script which would run the simple example in
reference checking mode for all reference traces in a given directory.
This script did not allow testing in parallel and its extensibility was
limited. The latter became more of an issue with the prospect of testing
against traces generated by our own encoder.

In order to accomodate for future extensions, we decided to replace the
existing script with a Makefile. Unfortunately our requirements for
individual jubs and some decisions lead to a Makefile which is a bit
more convoluted than we would have liked. However, so was the shell
script.